Fix warning message
git-svn-id: http://svn2.assembla.com/svn/secu3man@218 c1dae7de-a428-49f4-85d8-50b65f41d01d

+++ sources/MIdesk/MIDeskDlg.cpp
@@ -178,12 +178,12 @@ void CMIDeskDlg::SetUpdatePeriod(unsigned int i_period)
 
 void CMIDeskDlg::SetTachometerMax(int i_max)
 {
- m_tachometer.SetLimits(0, i_max);
+ m_tachometer.SetLimits(0, (float)i_max);
 }
 
 void CMIDeskDlg::SetPressureMax(int i_max)
 {
- m_pressure.SetLimits(10, i_max);
+ m_pressure.SetLimits(10, (float)i_max);
 }
 
 void CMIDeskDlg::Resize(const CRect& i_rect)
